requirements.txt
aiogram==3.4.1
firebase-admin==6.2.0
python-dotenv==1.0.0
openai>=1.0.0
